requirements.txt
pydantic==2.2.0
paho-mqtt==1.6.1
requests==2.31.0